Change cast for intpool to int64_t
The internal int type was changed to int64_t this
cast to int was missed.  Error found using the
Android compiler.

+++ util/oc_mmem.c
@@ -175,7 +175,7 @@ _oc_mmem_free(
       break;
     case INT_POOL:
       memmove(m->ptr, m->next->ptr,
-              &ints[OC_INTS_POOL_SIZE - avail_ints] - (int *)m->next->ptr);
+              &ints[OC_INTS_POOL_SIZE - avail_ints] - (int64_t *)m->next->ptr);
       break;
     case DOUBLE_POOL:
       memmove(m->ptr, m->next->ptr,
